Fixed FlightCsvLoader to correctly read the CSV using MAUI and implemented it into the services class.
