+++ GroupFlightProject/Data/Services.cs
@@ -6,16 +6,37 @@ using System.Threading.Tasks;
 
 namespace GroupFlightProject.Data
 {
-    public class Services
+	
+	public class Services
     {
         private List<Flight> flights;
-        public Services()
-        {
-            FlightCsvLoader flightLoader = new FlightCsvLoader();
-            flights = flightLoader.LoadFlights("GroupFlightProject/flights.csv").ToList();
-        }
+		private FlightCsvLoader flightLoader;
+		public Services()
+		{
+			flightLoader = new FlightCsvLoader();
+			LoadFlightsAsync();
+		}
 
-        public List<Flight> FindFlights(string departureLocation, string destination, string day)
+		internal class FlightCsvLoader
+		{
+			public List<Flight> LoadFlightsFromCsv(string csvContents)
+			{
+				var flights = new List<Flight>();
+				return flights;
+			}
+		}
+
+		private async Task LoadFlightsAsync()
+		{
+			using var stream = await FileSystem.OpenAppPackageFileAsync("flights.csv");
+			using var reader = new StreamReader(stream);
+
+			var csvContents = await reader.ReadToEndAsync();
+
+			flights = flightLoader.LoadFlightsFromCsv(csvContents);
+		}
+
+		public List<Flight> FindFlights(string departureLocation, string destination, string day)
         {
             return flights.Where(f => f.DepartureLocation.Equals(departureLocation, StringComparison.OrdinalIgnoreCase) &&
             f.Destination.Equals(destination, StringComparison.OrdinalIgnoreCase) &&
@@ -75,4 +96,39 @@ namespace GroupFlightProject.Data
             METHOD FOR MODIFYING RESERVATIONS
         }*/
     }
+	internal class FlightCsvLoader
+	{
+		public List<Flight> LoadFlightsFromCsv(string csvContents)
+		{
+			var flights = new List<Flight>();
+
+			string[] lines = csvContents.Split(Environment.NewLine, StringSplitOptions.RemoveEmptyEntries);
+
+			foreach (string line in lines)
+			{
+				var values = line.Split(',');
+
+				if (values.Length >= 8)
+				{
+					Flight newFlight = new Flight
+					{
+						FlightCode = values[0],
+						Airline = values[1],
+						DepartureLocation = values[2],
+						Destination = values[3],
+						Day = values[4],
+						Time = values[5],
+						AvailableSeats = int.Parse(values[6]),
+						FlightNumber = int.Parse(values[7])
+					};
+
+					flights.Add(newFlight);
+				}
+			}
+
+			return flights;
+		}
+	}
 }
+
+
